# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  BadRequestError:
    properties:
      error:
        example: Bad Request
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  EmailOTPRequest:
    properties:
      email:
        example: john.doe@gmail.com
        type: string
      otp:
        example: "123456"
        type: string
    required:
    - email
    - otp
    type: object
  EmailResendResponse:
    properties:
      message:
        example: We have sent the OTP to your Email address.
        type: string
      success:
        example: true
        type: boolean
    type: object
  InternalServerError:
    properties:
      error:
        example: Internal Server Error
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  LoginRequest:
    properties:
      password:
        example: "123"
        type: string
      username:
        example: vk@gmail.com
        type: string
    required:
    - password
    - username
    type: object
  LoginResponse:
    properties:
      expiry:
        example: "2025-05-01T12:00:00Z"
        type: string
      token:
        example: xxxxxxxxxxxxxxxxxxxxxxxxxxxx_adQssw5c
        type: string
      user_details:
        $ref: '#/definitions/UserResponse'
    type: object
  ResendEmailRequest:
    properties:
      email:
        example: john.doe@gmail.com
        type: string
    required:
    - email
    type: object
...
  Success-LoginResponse:
    properties:
      data:
        $ref: '#/definitions/LoginResponse'
      success:
        example: true
        type: boolean
    type: object
...
  Success-string:
    properties:
      data:
        type: string
      success:
        example: true
        type: boolean
    type: object
  UnauthorizedError:
    properties:
      error:
        example: Unauthorized
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  UserRegisterSuccess:
    properties:
      data:
        example: Please verify your Email Address. We have sent an OTP to the Email
          Address.
        type: string
      success:
        example: true
        type: boolean
    type: object
  UserRequest:
    properties:
      email:
        type: string
      first_name:
        maxLength: 50
        minLength: 2
        type: string
      last_name:
        maxLength: 50
        minLength: 2
        type: string
      phone:
        type: string
      role_id:
        type: string
    type: object
  UserResponse:
    properties:
      email:
        example: john.doe@gmail.com
        type: string
      first_name:
        example: John
        type: string
      full_name:
        example: John Doe
        type: string
      last_name:
        example: Doe
        type: string
      phone:
        example: "1234567890"
        type: string
      role_id:
        example: 97d699c0-24ff-48dc-b64a-c29353fa8865
        type: string
      user_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
...
  models.Role:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
      role_id:
        type: string
    type: object
  models.User:
    description: User model
    properties:
      email:
        type: string
      first_name:
        maxLength: 50
        minLength: 2
        type: string
      is_verified:
        type: boolean
      last_name:
        maxLength: 50
        minLength: 2
        type: string
      phone:
        type: string
      role:
        $ref: '#/definitions/models.Role'
      role_id:
        type: string
      user_id:
        type: string
      user_passwords:
        $ref: '#/definitions/models.UserPassword'
    required:
    - email
    - first_name
    - last_name
    - phone
    type: object
  models.UserPassword:
    properties:
      password:
        type: string
      user_id:
        type: string
      user_password_id:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: jaiswal.vaibhavkumar45@gmail.com
    name: API Support
  description: API for managing products and orders.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: E-Commerce API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user using email and password and returns a JWT
        token.
      parameters:
      - description: User login credentials
        in: body
        name: loginData
        required: true
        schema:
          $ref: '#/definitions/LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Authenticated successfully with JWT token
          schema:
            $ref: '#/definitions/Success-LoginResponse'
        "400":
          description: Invalid or malformed request body
          schema:
            $ref: '#/definitions/BadRequestError'
        "401":
          description: Invalid credentials or unauthorized access
          schema:
            $ref: '#/definitions/UnauthorizedError'
        "500":
          description: Unexpected server error
          schema:
            $ref: '#/definitions/InternalServerError'
      summary: User Login
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Registers a new user account by accepting valid email and other
        details.
      parameters:
      - description: User registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Registration successful
          schema:
            $ref: '#/definitions/UserRegisterSuccess'
        "400":
          description: Invalid input or missing required fields
          schema:
            $ref: '#/definitions/BadRequestError'
        "401":
          description: Unauthorized access attempt
          schema:
            $ref: '#/definitions/UnauthorizedError'
        "500":
          description: Unexpected server error
          schema:
            $ref: '#/definitions/InternalServerError'
      summary: Register a New User
      tags:
      - Auth
  /auth/resend-verification:
    post:
      consumes:
      - application/json
      description: Resends a verification code to the user's email address.
      parameters:
      - description: Email for which to resend OTP
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ResendEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OTP sent successfully
          schema:
            $ref: '#/definitions/EmailResendResponse'
        "400":
          description: Missing or invalid email
          schema:
            $ref: '#/definitions/BadRequestError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/InternalServerError'
      summary: Resend Verification Code
      tags:
      - Auth
  /auth/verify-email:
    post:
      consumes:
      - application/json
      description: Verifies a user's email address using an OTP sent to their email.
      parameters:
      - description: Email and OTP
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/EmailOTPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User verified successfully
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Missing or invalid OTP/email
          schema:
            $ref: '#/definitions/BadRequestError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/InternalServerError'
      summary: Verify Email with OTP
      tags:
      - Auth
...
  /logout:
    post:
      consumes:
      - application/json
      description: Invalidate the user's access token
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged out
          schema:
            $ref: '#/definitions/Success-string'
        "400":
          description: Invalid or malformed request body
          schema:
            $ref: '#/definitions/BadRequestError'
        "401":
          description: Invalid credentials or unauthorized access
          schema:
            $ref: '#/definitions/UnauthorizedError'
        "500":
          description: Unexpected server error
          schema:
            $ref: '#/definitions/InternalServerError'
      summary: Logout user
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"